+++ tests/bootstrap.php
@@ -24,3 +24,4 @@ define('NEGARESH_FILE', NEGARESH_PLUGIN_DIR . '/negaresh.php');
 require_once NEGARESH_PLUGIN_DIR . '/includes/Virastar.php';
 require_once NEGARESH_PLUGIN_DIR . '/includes/negaresh-settings.php';
 require_once NEGARESH_PLUGIN_DIR . '/includes/negaresh-class.php';
+require_once NEGARESH_PLUGIN_DIR . '/includes/negaresh-editor.php';